+++ front-end/src/components/WorkflowInput/WorkflowInput.tsx
@@ -228,6 +228,20 @@ class WorkflowInput extends React.Component<WorkflowInputProps, WorkflowInputSta
   };
 
   /**
+   * Clear all set input data.
+   */
+  clearInput = () => {
+    this.setState({ inputs: [] });
+  };
+
+  /**
+   * Clear all set output data.
+   */
+  clearOutput = () => {
+    this.setState({ outputs: [] });
+  };
+
+  /**
    * Change the output at the given index to the edited data.
    * @param index - index of the data in state.inputs.
    * @param newOutput - the edited data.
@@ -279,6 +293,13 @@ class WorkflowInput extends React.Component<WorkflowInputProps, WorkflowInputSta
     const { constraints } = this.state;
     constraints.splice(index, 1);
     this.setState({ constraints });
+  };
+
+  /**
+   * Clear all set constraints.
+   */
+  clearConstraints = () => {
+    this.setState({ constraints: [] });
   };
 
   /**
@@ -952,6 +973,7 @@ class WorkflowInput extends React.Component<WorkflowInputProps, WorkflowInputSta
             onRemove={this.onInputRemove}
             inOuts={inputs}
             tooltip="The input(s) you want your workflow to have."
+            clearWorkflowData={this.clearInput}
           />
           {/* Output */}
           <InOutBox
@@ -963,6 +985,7 @@ class WorkflowInput extends React.Component<WorkflowInputProps, WorkflowInputSta
             onRemove={this.onOutputRemove}
             inOuts={outputs}
             tooltip="The output(s) you want your workflow to have."
+            clearWorkflowData={this.clearOutput}
           />
           {/* Constraints */}
           <ConstraintBox
@@ -981,6 +1004,7 @@ class WorkflowInput extends React.Component<WorkflowInputProps, WorkflowInputSta
             defaultData={this.defaultData}
             defaultTool={this.defaultTool}
             defaultConstraint={this.defaultConstraint}
+            clearConstraints={this.clearConstraints}
           />
           {/* Run */}
           <WorkflowRun