+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2021-12-20
+
+### Added
+
+- Clear buttons to clear all input data, output data, constraints on the explore page.
+- Domains can now be permanently deleted in the domain edit page.
+
+### Changed
+
+- Update back-end Kotlin version from 1.3.72 -> 1.6.0.
+- Update APE dependency version 1.1.9 -> 1.1.12 (includes log4j vulnerability fix).
+
+### Fixed
+
+- Remove "tab" as suggestion for separating data taxonomy roots as this also moves the user to the next input field. 
+- Some dependencies were updated to fix vulnerability issues.
+- When there are many inputs, outputs, or constraints on the explore page; a scrollbar will now appear inside the workflow input box to prevent the workflow result from appearing far below.
+
 ## [1.4.0] - 2021-11-06
 
 ### Added
@@ -157,6 +175,7 @@ This release is paired with the APE Web [DOI release](https://zenodo.org/badge/l
   * Approve user accounts.
   
 [Unreleased]: https://github.com/sanctuuary/APE-Web/compare/master...dev
+[1.5.0]: https://github.com/sanctuuary/APE-Web/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/sanctuuary/APE-Web/compare/v1.3.3...v1.4.0
 [1.3.3]: https://github.com/sanctuuary/APE-Web/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/sanctuuary/APE-Web/compare/v1.3.1...v1.3.2